+++ TOMB4/game/deltapak.h
@@ -95,6 +95,10 @@ void special3_end();
 void special3_control();
 void special1_init();
 void special1_end();
+#ifdef TIMES_LEVEL
+void times_init();
+void times_control();
+#endif
 
 extern long cutseq_trig;
 extern long cutseq_num;